[v2] Drop all legacy filter artwork fields in favour of new connection.
