+++ patches/@types/graphql-relay+0.4.9.patch
@@ -1,15 +1,26 @@
 patch-package
 --- a/node_modules/@types/graphql-relay/index.d.ts
 +++ b/node_modules/@types/graphql-relay/index.d.ts
-@@ -72,6 +72,7 @@ export interface ConnectionConfig {
+@@ -71,7 +71,9 @@ export interface ConnectionConfig {
+     resolveNode?: GraphQLFieldResolver<any, any> | null;
      resolveCursor?: GraphQLFieldResolver<any, any> | null;
      edgeFields?: Thunk<GraphQLFieldConfigMap<any, any>> | null;
++    edgeInterfaces?: GraphQLInterfaceType[];
      connectionFields?: Thunk<GraphQLFieldConfigMap<any, any>> | null;
-+    interfaces?: [GraphQLInterfaceType];
++    connectionInterfaces?: GraphQLInterfaceType[];
  }
  
  export interface GraphQLConnectionDefinitions {
-@@ -211,11 +212,11 @@ export function getOffsetWithDefault(
+@@ -104,6 +106,8 @@ export interface PageInfo {
+     hasNextPage?: boolean | null;
+ }
+ 
++export const PageInfoType: GraphQLObjectType<any, any>
++
+ /**
+  * A flow type designed to be exposed as a `Connection` over GraphQL.
+  */
+@@ -211,11 +215,11 @@ export function getOffsetWithDefault(
  
  // mutation/mutation.js
  
@@ -25,7 +36,7 @@ patch-package
  
  /**
   * A description of a mutation consumable by mutationWithClientMutationId
-@@ -231,12 +232,12 @@ export type mutationFn = (
+@@ -231,12 +235,12 @@ export type mutationFn = (
   * input field, and it should return an Object with a key for each
   * output field. It may return synchronously, or return a Promise.
   */
@@ -41,7 +52,7 @@ patch-package
      deprecationReason?: string;
  }
  
-@@ -244,8 +245,8 @@ export interface MutationConfig {
+@@ -244,8 +248,8 @@ export interface MutationConfig {
   * Returns a GraphQLFieldConfig for the mutation described by the
   * provided MutationConfig.
   */
