+++ patches/graphql-relay+0.5.4.patch
@@ -31,11 +31,52 @@ patch-package
    var firstEdge = edges[0];
 --- a/node_modules/graphql-relay/lib/connection/connection.js
 +++ b/node_modules/graphql-relay/lib/connection/connection.js
-@@ -88,6 +88,7 @@ function connectionDefinitions(config) {
+@@ -3,7 +3,7 @@
+ Object.defineProperty(exports, "__esModule", {
+   value: true
+ });
+-exports.connectionArgs = exports.backwardConnectionArgs = exports.forwardConnectionArgs = undefined;
++exports.connectionArgs = exports.backwardConnectionArgs = exports.forwardConnectionArgs = exports.PageInfoType = undefined;
+ 
+ var _extends = Object.assign || function (target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i]; for (var key in source) { if (Object.prototype.hasOwnProperty.call(source, key)) { target[key] = source[key]; } } } return target; }; /**
+                                                                                                                                                                                                                                                                    * Copyright (c) 2015-present, Facebook, Inc.
+@@ -69,6 +69,7 @@ function connectionDefinitions(config) {
+   var edgeType = new _graphql.GraphQLObjectType({
+     name: name + 'Edge',
+     description: 'An edge in a connection.',
++    interfaces: config.edgeInterfaces,
+     fields: function fields() {
+       return _extends({
+         node: {
+@@ -88,6 +89,7 @@ function connectionDefinitions(config) {
    var connectionType = new _graphql.GraphQLObjectType({
      name: name + 'Connection',
      description: 'A connection to a list of items.',
-+    interfaces: config.interfaces,
++    interfaces: config.connectionInterfaces,
      fields: function fields() {
        return _extends({
          pageInfo: {
+@@ -108,7 +110,7 @@ function connectionDefinitions(config) {
+ /**
+  * The common page info type used by all connections.
+  */
+-var pageInfoType = new _graphql.GraphQLObjectType({
++var pageInfoType = exports.PageInfoType = new _graphql.GraphQLObjectType({
+   name: 'PageInfo',
+   description: 'Information about pagination in a connection.',
+   fields: function fields() {
+--- a/node_modules/graphql-relay/lib/index.js
++++ b/node_modules/graphql-relay/lib/index.js
+@@ -30,6 +30,12 @@ Object.defineProperty(exports, 'forwardConnectionArgs', {
+     return _connection.forwardConnectionArgs;
+   }
+ });
++Object.defineProperty(exports, 'PageInfoType', {
++  enumerable: true,
++  get: function get() {
++    return _connection.PageInfoType;
++  }
++});
+ 
+ var _arrayconnection = require('./connection/arrayconnection.js');
+
