+++ src/schema/v1/index.ts
@@ -14,7 +14,7 @@ if (enableSchemaStitching) {
     if (typeof jest == "undefined") {
       console.warn("[FEATURE] Enabling Schema Stitching")
     }
-    exportedSchema = incrementalMergeSchemas(exportedSchema)
+    exportedSchema = incrementalMergeSchemas(exportedSchema, 1)
   } catch (err) {
     console.log("Error merging schemas:", err)
   }
