+++ src/schema/v2/tag.ts
@@ -8,7 +8,6 @@ import {
   GraphQLNonNull,
   GraphQLFieldConfig,
 } from "graphql"
-import filterArtworks from "./filter_artworks"
 import { queriedForFieldsOtherThanBlacklisted } from "lib/helpers"
 import { ResolverContext } from "types/graphql"
 
@@ -33,7 +32,6 @@ export const TagType = new GraphQLObjectType<any, ResolverContext>({
       count: {
         type: GraphQLInt,
       },
-      filteredArtworks: filterArtworks("tag_id"),
     }
   },
 })
